Extract shared page view model cache
